Implementing Stephanie's new idea to calculate epsilon
Analyzing Rob Brewster's data, but now also incorporating the fact that
the Lac operator copy number influences how many repressors will be in
the active state (the more operators present, the more repressors will
be in the active state).
This also has a bearing on what the DNA binding energies that Hernan
measured represent.
The conclusion seems to be that epsilon can take on any positive value,
and this value is not well fixed by the fitting, which mean's that
Hernan's approximation (that all repressors are active) should not
hamper the data fitting. Ultimately, we will probably need to just
choose an epsilon value and stick with it, and declare in the SI that
there is uncertainty in this value