Added Manuel's thought experiment on removing tetramerization region
What happens if we remove the tetramerization region in the Lac
repressor and then measure fold-change? It turns out that you can
determine whether the two repressor dimers are allosterically
independent or not. Furthermore, if the two dimers are not
allosterically independent, this gives you a potentially new way to
measure epsilon. What a bonanza!